+++ src/main/java/com/andersen/hw/service/TicketService.java
@@ -7,15 +7,15 @@ import java.util.List;
 public interface TicketService {
     void addTicket(Ticket ticket);
 
-    Ticket getById(Long id);
+    Ticket getById(Integer id);
 
     List<Ticket> getAll();
 
-    void deleteById(Long id);
+    void deleteById(Integer id);
 
     void updateTicket(Ticket ticket);
 
-    void share(Long ticketId, String phone);
+    void share(Integer ticketId, String phone);
 
-    void share(Long ticketId, String phone, String email);
+    void share(Integer ticketId, String phone, String email);
 }
